% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{convertToFactors}
\alias{convertToFactors}
\title{convertToFactors()}
\usage{
convertToFactors(x)
}
\arguments{
\item{x}{(df)}
}
\value{
A df with all character columns as factors
}
\description{
convert all character columns of a df to factors
}
