% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{densPlotMultiLines}
\alias{densPlotMultiLines}
\title{densPlotMultiLines()}
\usage{
densPlotMultiLines(s, main = "", xlab = "")
}
\arguments{
\item{s}{A matrix}

\item{main}{main title of the plot}

\item{xlab}{label for x}
}
\value{
A denstiy plot of multiple lines, each line is colored (represent values in a column)
}
\description{
Plot density of a matrix, each line represent a column
}
