% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{excludeMatch}
\alias{excludeMatch}
\title{excludeMatch()}
\usage{
excludeMatch(df, column, value)
}
\arguments{
\item{df}{A data.frame to be filtered}

\item{column}{(str) column to be filtered}

\item{value}{(str) a str to match with the value column, match of the value is removed.}
}
\value{
A data.frame with matched records removed, including empty df if all records matched
}
\description{
Table function: remove rows of column with exact match of a specified value
}
\note{
only exact match is removed
}
\examples{
column <- "date"
value <- "2014-01-01" # remove records with date 2014-01-01
excludeMatch(df, column, value)
}
