% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{filterContain}
\alias{filterContain}
\title{filterContain()}
\usage{
filterContain(df, column, value, verbose = F, ...)
}
\arguments{
\item{df}{A data.frame to be filtered}

\item{column}{(str) column to be filtered}

\item{value}{(str list) a list of strs to match with the value column. If a match is found, the record is returned}

\item{verbose}{[optional] (bool) print if a match is found, and how many records returned. default = F}

\item{...}{more parameters in the grepl() function, eg. ignore.case = T}
}
\value{
A data.frame with matched records, including empty df if no match is found
}
\description{
Table function: filter a df by column contains a list of specified values
}
\note{
only to match the full words in the list (if the value contains two or more, it is a match)
}
\examples{
filterContain(df, column, value, verbose = F, ...)

 filter a df by column contains a list of specified values
 only to match the full words in the list (if the value contains two or more, will match)
 ... more parameters in the grepl() function, eg. ignore.case = T
}
