% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{filterLess}
\alias{filterLess}
\title{filterLess()}
\usage{
filterLess(df, column, value)
}
\arguments{
\item{df}{A data.frame to be filtered}

\item{column}{(str) the date column to be filtered}

\item{value}{(date) date in format yyyy-dd-mm}
}
\value{
A data.frame with record date before or equal to the specified date, including empty df if no match is found
}
\description{
Table function: filter a df by a specified date, return df with dates prior or on the date
}
\examples{
column <- "date"
value <- "2014-01-01"
filterLess(df, column, value)
}
