% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{filterMatch}
\alias{filterMatch}
\title{filterMatch()}
\usage{
filterMatch(df, column, value)
}
\arguments{
\item{df}{A data.frame to be filtered}

\item{column}{(str) column to be filtered}

\item{value}{(str) a str to match with the value column.}
}
\value{
A data.frame with matched records, including empty df if no match is found
}
\description{
Table function: filter a df by column with exact match of specified values
}
\note{
only exact match is returned
}
\examples{
column <- "date"
value <- "2014-01-01"
filterMatch(df, column, value)

}
