% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ggColorHue}
\alias{ggColorHue}
\title{ggColorHue()}
\usage{
ggColorHue(n)
}
\arguments{
\item{n}{numeric, number of colors for output}
}
\value{
a list of colors simulate ggplot default color
}
\description{
return a list of colors simulate ggplot default color
}
