% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{mapBindAllColumns}
\alias{mapBindAllColumns}
\title{mapBindAllColumns()}
\usage{
mapBindAllColumns(x, y, id_col = "")
}
\arguments{
\item{x}{(df), can be NULL}

\item{y}{(df), can be NULL}

\item{id_col}{[optional] (str) a column that are shared between x and y and must have unique values within each df. 
If not specified, the row.names of x and y are used for the match}
}
\value{
A dataframe that either 1. match the row.names of x and y, and combine the columns from x and y (id_col='')
2. match the content of id_col, and combine the rest of the columns of x and y
}
\description{
Table Functions: rbind two data frames x, y and include all columns from both data frame
}
\note{
If row.names are not meaningful (default as numerical index) 
       and id_col is not specified, the combined df is not meaningful 
       because the rows are combined based on the numerical order
If a column name in y is also in x (except for the id_col), the column name in y is suffixed "_2"
}
\examples{
mapBindAllColumns(x, y, id_col = "match_col")
}
