% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{mapValue}
\alias{mapValue}
\title{mapValue()}
\usage{
mapValue(x, y, verbose = F)
}
\arguments{
\item{x}{A list with values to be mapped}

\item{y}{A named list, where names(y) correponding to the value in x, and values}
}
\value{
A named list of the new mapped values
}
\description{
Given a list x, and a named list of y. Map the values of y to x based on the names of y.
}
\note{
The named list of y must contain all values of x in names(y). The returned list is the same order as x.
}
\examples{
x <- c(rep("WT", 3), rep("Rasgrf1_KO", 3))
y <- c(Rasgrf1_KO = "#2", WT = "#1", WT2 = "#3")
mapValue(x,y)
}
