% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{multiplot()}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL, order = T,
  sequence = NULL)
}
\arguments{
\item{...}{a list of ggplot objects, e.g. p1, p2}

\item{plotlist}{a list of ggplot objects, e.g. list(p1, p2)}

\item{cols}{(integer) specify how many columns in the output plot}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored. e.g matrix(c(1,2,3,4), nrow=2, byrow=TRUE)}

\item{order}{(bool) default T
if T, the order is from top to bottom, then left to right
if F, the order is from left to right then top to bottom}
}
\value{
a plot with all the input plots
}
\description{
plot multiple ggplots in one plot
Multiple plot function # from http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
\note{
If the layout is something like matrix(c(1,2,3,4), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.The same plot can be repeated
}
