% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{orderCol}
\alias{orderCol}
\title{orderCol()}
\usage{
orderCol(df, cols)
}
\arguments{
\item{df}{input df}

\item{cols}{column names in strings (add - if want to sort by descending order()only for numeric )}
}
\description{
reorder df by specified columns
}
\examples{
df <- data.frame(col1 = c('a', 'a', 'b','b'),
                 col2 = c('orange', 'red', 'red', 'grey'),
                 col3 = c(3,6,8,5))
cols = c('col1', 'col2')
orderCol(df, cols)
orderCol(df, '-col3')
}
