% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{plotBox}
\alias{plotBox}
\title{plotBox()}
\usage{
plotBox(df, x, y, palette_defined = F, palette = "", title = "",
  log_scale = F, x_angle = 30, text_size = 8, ...)
}
\arguments{
\item{df}{A data.frame}

\item{x}{(str) column to plot x (column name), fill colour by x}

\item{y}{(str) column to plot y (column name)}

\item{palette_defined}{(bool) indicate whether to use as named palette list as defined in palette.
default = F: use ggplot default colour scheme}

\item{palette}{[optional] (named str) define the colour of the factor levels}

\item{title}{[optional] (str) title of the plot, default none}

\item{log_scale}{(bool) If True, log 10 scale y axis. default = F}

\item{x_angle}{[optional] (num) angle of x axis labels, default = 30}

\item{text_size}{(num) plot title size, default = 8}

\item{...}{other param in the ggplot2::geom_boxplot()}
}
\value{
A ggplot boxplot
}
\description{
Boxplot (ggplot style)
}
\note{
use palette to define color for the factors. The palette list is a named list of colors. The names are corresponding to the factor levels
    e.g. palette <- c(red = "#ff0000", blue="#333BFF", pink ="#DD1C77", orange = "#CC6600", purple ="#9633FF",l_blue = "#00ffff", l_green = "#9bbc57", yellow = "#ffe700", salmon = "#ff7856",l_purple = "#e47fff")
}
