% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{plotXY}
\alias{plotXY}
\title{plotXY()}
\usage{
plotXY(df, x, y, palette_defined = F, palette = "", group = "",
  title = "", facet_gp = "", facet_grid = "", x_angle = 45,
  text_size = 8, log_x = F, log_y = F, plt_alpha = 0.4, ...)
}
\arguments{
\item{df}{A data.frame}

\item{x}{(str) column to plot x (column name)}

\item{y}{(str) column to plot y (column name)}

\item{palette_defined}{(bool) indicate whether to use as named palette list as defined in palette.
default = F: use ggplot default colour scheme}

\item{palette}{[optional] (named str) define the colour of the factor levels}

\item{group}{[optional] (str) fill colour by a column (column name), default is do not group by any column}

\item{title}{[optional] (str) title of the plot, default none}

\item{facet_gp}{[optional] (str) the column name for the facet group (as in facet_wrap())}

\item{x_angle}{[optional] (num) angle of x axis labels, default = 45}

\item{text_size}{(num) plot title size, default = 8}

\item{log_x}{(bool) If True, log 10 scale x axis. default = F}

\item{log_y}{(bool) If True, log 10 scale y axis. default = F}

\item{plt_alpha}{[optional] (num) plot alpha (transparency), default = 0.4}

\item{...}{other param in the ggplot2::geom_point()}
}
\value{
A ggplot XY plot
}
\description{
plot XY (ggplot style)
}
\note{
use palette to define color for the factors. The palette list is a named list of colors. The names are corresponding to the factor levels
    e.g. palette <- c(red = "#ff0000", blue="#333BFF", pink ="#DD1C77", orange = "#CC6600", purple ="#9633FF",l_blue = "#00ffff", l_green = "#9bbc57", yellow = "#ffe700", salmon = "#ff7856",l_purple = "#e47fff")
}
