% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rbindAllColumns}
\alias{rbindAllColumns}
\title{rbindAllColumns()}
\usage{
rbindAllColumns(x, y)
}
\arguments{
\item{x}{(df), if NULL, then return the other df.}

\item{y}{(df), if NULL, then return the other df.}
}
\value{
a combined df with all the columns from x and y.
}
\description{
Table Functions: rbind two data frames x, y and include all columns from both data frame
}
\note{
If one of the input is null, then return the non-NULL df. 
No mapping of the same rows. All rows are added. NA to missing values.
If mapping is needed, use mapBindAllColumns()
}
\examples{
rbindAllColumns(x, y)
}
