% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rbindDF}
\alias{rbindDF}
\title{rbindDF()}
\usage{
rbindDF(df1, df2, ...)
}
\arguments{
\item{...}{more input data.frames}

\item{df1:}{input data.frame 1}

\item{df2:}{input data.frame 2}
}
\description{
rbind rows of multiple dataframes, if one or more of the input is not class data.frame. these inputs will be ignored.
only data.frames will be rbind.
}
\examples{
rbindDF(df1,df2,df3,...)
}
