% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vennDiagram2}
\alias{vennDiagram2}
\title{vennDiagram2()}
\usage{
vennDiagram2(input_ls, title = "", col_palette = RColorBrewer::brewer.pal(9,
  "Set1"), ...)
}
\arguments{
\item{input_ls}{A list, e.g. input_ls = list(list1 = c("A", "B", "C"), list2= c("A", "B", "D"))}

\item{title}{(str) plot title}

\item{...}{more arguments in VennDiagram::venn.diagram()}

\item{palette}{(list of named colors) default is brewer.pal(8,"Accent")}
}
\value{
a venn diagram
}
\description{
Wrapper for VennDiagram::venn.diagram() function
}
\note{
the plot will overlap with the old plot, use graphics.off() to clear old plots
not recomended for over 5 lists (very messy with more than 5 lists)
}
\examples{
graphics.off()
input_ls = list(list1 = c("A", "B", "C"), list2= c("A", "B", "D"))
pal <- c(list1 = "red", list2 = "blue")  ## if the colour is named the same as the list name, the colour will correponding to the specific list
vennDiagram2(input_ls, col_palette=pal)

graphics.off()
vennDiagram2(input_ls)

}
